Refactor : refactor directory
